Fix ret_dom option bug with upper and mixed case tags.
If we use ret_dom option, a first tag in out_str is searched in lower
case. This cause a problem if a template is commented and tag names in
upper or mixed case.

Fix this by use toLowerCase() method on this.out_str before searching a
first tag name.