process attribute value on _parse_tag_attrs() acll
Use for attribute value parsing a callback specified in a tag spec.
If some attribute value should be parsed(i.e. convert var.long.name to
an array), then a function for this work should be specified in a
tag spec. It must accept 2 arguments - a tag spec and a tag attribute
value to be parsed. It must return an attribute parsed value.